kernel: move sched_priq.h to internal/ folder
This header is internal to the kernel and shall not be included directly.
