+++ include/zephyr/kernel/internal/sched_priq.h
