+++ include/zephyr/kernel_includes.h
@@ -24,7 +24,7 @@
 #include <zephyr/linker/sections.h>
 #include <zephyr/sys/atomic.h>
 #include <zephyr/sys/__assert.h>
-#include <zephyr/kernel/sched_priq.h>
+#include <zephyr/kernel/internal/sched_priq.h>
 #include <zephyr/sys/dlist.h>
 #include <zephyr/sys/slist.h>
 #include <zephyr/sys/sflist.h>
