+++ kernel/include/wait_q.h
@@ -12,7 +12,7 @@
 #include <zephyr/kernel_structs.h>
 #include <zephyr/sys/dlist.h>
 #include <zephyr/sys/rb.h>
-#include <zephyr/kernel/sched_priq.h>
+#include <zephyr/kernel/internal/sched_priq.h>
 #include <timeout_q.h>
 
 #ifdef __cplusplus
